hibernate

+++ 18_jpa_hibernate/hibernate_part_two/src/main/java/com/nixsolutions/servlet/OrderPartPostServlet.java
@@ -68,10 +68,10 @@ public class OrderPartPostServlet extends HttpServlet {
 				if (part_id != null) {
 					OrderPart orderPart = orderPartDao.getByPK(Integer.parseInt(order_id), Integer.parseInt(part_id));
 					if (orderPart == null) {
-						orderPart = new OrderPart(orderDao.getByPK(Integer.parseInt(order_id)), partDao.getByPK(Integer.parseInt(part_id)), Integer.parseInt(used_amount));
+						orderPart = new OrderPart(orderDao.getByPK(Long.parseLong(order_id)), partDao.getByPK(Long.parseLong(part_id)), Long.parseLong(used_amount));
 						orderPartDao.createFrom(orderPart);
 					} else {
-						orderPart.setUsedAmount(Integer.parseInt(used_amount));
+						orderPart.setUsedAmount(Long.parseLong(used_amount));
 						orderPartDao.update(orderPart);
 					}					
 				}
